Feature/front end (#23)
* fix webapi actions

* edit table

* admin panel (views:tables and dishes)

* create new project

* Update README.md

* CRM tables and dishes (#12)

* Add identity in Identity.Project (#13)

* CRM tables and dishes

* Login and register in Identity project

* Update identity (#14)

* CRM tables and dishes

* Login and register in Identity project

* Update identity 1.02

* Feature/identity server (#15)

* identity project

* Some front

* Update README.md

* Identity/test2 (#16)

* identity

* Credentials IDS4

* Configure and mvc identityserver4

* Register and login IdentityServer4

* Js menu

* id_token, user claims access_token

* logOut, make some front

* add roles IdentityServer4

* localize main page

* localize ui

* Localization and basket

* Access token for WebAPI

* localize

* crm

* basket page, js basket, order dish and view on basket page

* bind user to table

* bind user to table

* fix bug with edit and delete on table/tables

* updating pages js

* fix bug basket and table

* fix bug with quantity and delete quantity dish of basket

* fix bug with invalid login credential

* notify for waiter about dish, unit test(not completed)

* unit tests not working (nullException)

* unit tests

* order migration, unit test of order, fix signalR bug with output data, CRUD of orders

* add feature - select orders for id user

* frontend of login

* fix bug with delete operation

* some front on menu

* Basket delete at compeleted page, create link at js constant

+++ Restraunt.Data/Repositories/TableRepository.cs
@@ -22,46 +22,80 @@ namespace Restraunt.Data.Repositories
         }
         public async Task<bool> Create(TableDto? entity)
         {
-           var result = new Table { Id = entity.Id, Name = entity.Name, Link = entity.Link};
+
+            var result = new Table { Id = entity.Id, Name = entity.Name, Link = entity.Link };
 
             await _db.Tables.AddAsync(result);
-            
+
+
             return true;
         }
 
         public async Task<bool> Delete(Guid Id)
         {
             var entity = await _db.Tables.Where(t => t.Id == Id).FirstOrDefaultAsync();
-            _db.Tables.Remove(entity);
+            if (entity != null)
+            {
+                _db.Tables.Remove(entity);
+                return true;
+            }
+            else
+            {
+                return false;
+            }
 
-            
-            return true;
         }
 
         public async Task<Table> Get(Guid id)
         {
+            var table = await _db.Tables.Where(t => t.Id == id).Include(u => u.User).FirstOrDefaultAsync();
 
-            var table = await _db.Tables.Where(t => t.Id == id).FirstOrDefaultAsync();
+
 
             return table;
         }
 
         public async Task<IEnumerable<Table>> Select()
         {
-            return _db.Tables.ToList();
+
+            return _db.Tables.Include(u => u.User).ToList();
         }
 
-        public async Task<Table> Edit(TableDto entity)
+        public async Task<Table> Edit(EditTableDto entity)
         {
             var table = await _db.Tables.Where(t => t.Id == entity.Id).FirstOrDefaultAsync();
 
-            if(table != null)
+            if (table != null)
             {
-                table.Name = entity.Name; table.User = entity.User;
+                table.Name = entity.Name;
+
+                _db.Update(table);
+
+            }
+            return table;
+
+        }
+
+        public async Task<Table> BindUserToTable(BindUserToTableDto model)
+        {
+            Table? table = await _db.Tables.Where(t => t.Id == model.Id).FirstOrDefaultAsync();
+
+            User? user = await _db.Users.FirstOrDefaultAsync(u => u.Id == model.UserId);
+
+            if (table != null)
+            {
+                table.User = user;
+
+
 
                 _db.Update(table);
             }
             return table;
+
         }
+
+
+      
+
     }
 }
